+++ lib/src/components/components.dart
@@ -22,3 +22,4 @@ export 'input_field.dart';
 export 'checkmark.dart';
 export 'segmented_control.dart';
 export 'gradient_widget.dart';
+export 'message_box.dart';